backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
cryptography==41.0.3
Werkzeug==3.0.1
python-dotenv==1.0.0
pylint==2.17.4
PyJWT==2.8.0